IGNITE-19786 Move Persistence and Property task classes to command class packages (#10794)
