+++ modules/core/src/main/java/org/apache/ignite/internal/management/property/PropertyGetCommand.java
@@ -18,8 +18,6 @@
 package org.apache.ignite.internal.management.property;
 
 import java.util.function.Consumer;
-import org.apache.ignite.internal.commandline.property.tasks.PropertyOperationResult;
-import org.apache.ignite.internal.commandline.property.tasks.PropertyTask;
 import org.apache.ignite.internal.management.api.ComputeCommand;
 
 /** */
