+++ modules/core/src/main/resources/META-INF/classnames.properties
@@ -377,12 +377,12 @@ org.apache.ignite.internal.management.meta.MetadataRemoveTask$DropAllThinSession
 org.apache.ignite.internal.management.meta.MetadataRemoveTask$MetadataRemoveJob
 org.apache.ignite.internal.management.meta.MetadataUpdateTask
 org.apache.ignite.internal.management.meta.MetadataUpdateTask$MetadataUpdateJob
-org.apache.ignite.internal.commandline.property.tasks.PropertiesListResult
-org.apache.ignite.internal.commandline.property.tasks.PropertiesListTask
-org.apache.ignite.internal.commandline.property.tasks.PropertiesListTask$PropertiesListJob
-org.apache.ignite.internal.commandline.property.tasks.PropertyOperationResult
-org.apache.ignite.internal.commandline.property.tasks.PropertyTask
-org.apache.ignite.internal.commandline.property.tasks.PropertyTask$PropertyJob
+org.apache.ignite.internal.management.property.PropertiesListResult
+org.apache.ignite.internal.management.property.PropertiesListTask
+org.apache.ignite.internal.management.property.PropertiesListTask$PropertiesListJob
+org.apache.ignite.internal.management.property.PropertyOperationResult
+org.apache.ignite.internal.management.property.PropertyTask
+org.apache.ignite.internal.management.property.PropertyTask$PropertyJob
 org.apache.ignite.internal.compute.ComputeTaskCancelledCheckedException
 org.apache.ignite.internal.compute.ComputeTaskTimeoutCheckedException
 org.apache.ignite.internal.direct.DirectMessageReader$1
@@ -2248,9 +2248,9 @@ org.apache.ignite.internal.visor.misc.VisorWalTaskResult
 org.apache.ignite.internal.visor.performancestatistics.VisorPerformanceStatisticsOperation
 org.apache.ignite.internal.management.performancestatistics.PerformanceStatisticsTask
 org.apache.ignite.internal.management.performancestatistics.PerformanceStatisticsTask$VisorPerformanceStatisticsJob
-org.apache.ignite.internal.visor.persistence.PersistenceTask
-org.apache.ignite.internal.visor.persistence.PersistenceTask$PersistenceJob
-org.apache.ignite.internal.visor.persistence.PersistenceTaskResult
+org.apache.ignite.internal.management.persistence.PersistenceTask
+org.apache.ignite.internal.management.persistence.PersistenceTask$PersistenceJob
+org.apache.ignite.internal.management.persistence.PersistenceTaskResult
 org.apache.ignite.internal.management.kill.ContinuousQueryCancelTask
 org.apache.ignite.internal.management.kill.ContinuousQueryCancelTask$ContinuousQueryCancelJob
 org.apache.ignite.internal.management.kill.QueryCancelOnInitiatorTask
